CAKE-272 | moved constants into localization_constants.dart and locale_list.dart; moved generation of localized strings inside of the loop body
