+++ tool/localization/locale_list.dart
@@ -1,0 +1,3 @@
+const locales = ['en', 'de', 'es', 'hi',
+                 'ja', 'ko', 'nl', 'pl',
+                 'pt', 'ru', 'uk', 'zh'];
